Calculando o preenchimento do salao na vertical (O Salão do Clube)

+++ O Salão do Clube/O Salão do Clube/Entities/Salao.cs
@@ -9,6 +9,6 @@ namespace O_Salão_do_Clube.Entities
     static class Salao
     {
         public static int Comprimento { get; set; }
-        public static int Largura { get; set; }
+        public static double Largura { get; set; }
     }
 }
